rename: 파일 구조 변경 및 이름 변경
