+++ app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import StyledComponentsRegistry from "@/lib/styleRegistry";
 import "./globals.css";
-import Navigation from "@/components/navigation";
+import Navigation from "@/components/common/navigation";
 import Layout from "@/components/layout";
 
 export const metadata: Metadata = {
