+++ app/tv/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import ContentList from "@/components/contentList";
+import ContentList from "@/components/pages/contents/contentList";
 
 export default function TVPage() {
   return <ContentList contentType="tv" />;
