+++ components/common/customSlider.tsx
