+++ components/pages/searchResults/companyList.tsx
